requirements.txt
streamlit
fastapi
uvicorn
requests
beautifulsoup4
pandas
lxml
python-dotenv
google-genai
sqlalchemy
tqdm
